Simplify event handling by removing EventManager and putting event handler directly in objects
Also aligns with standard .NET API designs
