+++ LibVLCSharp/Shared/MediaPlayer.cs
@@ -630,8 +630,6 @@ namespace LibVLCSharp.Shared
             internal static extern void LibVLCMediaPlayerSetAndroidContext(IntPtr mediaPlayer, IntPtr aWindow);
 #endif
         }
-        
-        MediaPlayerEventManager _eventManager;
 
         /// <summary>Create an empty Media Player object</summary>
         /// <param name="libVLC">
@@ -640,7 +638,7 @@ namespace LibVLCSharp.Shared
         /// </param>
         /// <returns>a new media player object, or NULL on error.</returns>
         public MediaPlayer(LibVLC libVLC) 
-            : base(() => Native.LibVLCMediaPlayerNew(libVLC.NativeReference), Native.LibVLCMediaPlayerRelease)
+            : base(() => Native.LibVLCMediaPlayerNew(libVLC.NativeReference), Native.LibVLCMediaPlayerRelease, Native.LibVLCMediaPlayerEventManager)
         {
         }
 
@@ -651,7 +649,7 @@ namespace LibVLCSharp.Shared
         /// </param>
         /// <returns>a new media player object, or NULL on error.</returns>
         public MediaPlayer(Media media)
-            : base(() => Native.LibVLCMediaPlayerNewFromMedia(media.NativeReference), Native.LibVLCMediaPlayerRelease)
+            : base(() => Native.LibVLCMediaPlayerNewFromMedia(media.NativeReference), Native.LibVLCMediaPlayerRelease, Native.LibVLCMediaPlayerEventManager)
         {
         }
         
@@ -668,22 +666,6 @@ namespace LibVLCSharp.Shared
                 return mediaPtr == IntPtr.Zero ? null : new Media(mediaPtr);
             }
             set => Native.LibVLCMediaPlayerSetMedia(NativeReference, value.NativeReference);
-        }
-
-        /// <summary>
-        /// Get the Event Manager from which the media player send event.
-        /// </summary>
-        public MediaPlayerEventManager EventManager
-        {
-            get
-            {
-                if (_eventManager == null)
-                {
-                    var eventManagerPtr = Native.LibVLCMediaPlayerEventManager(NativeReference);
-                    _eventManager = new MediaPlayerEventManager(eventManagerPtr);
-                }
-                return _eventManager;
-            }
         }
 
         /// <summary>
@@ -1617,11 +1599,819 @@ namespace LibVLCSharp.Shared
         public bool SetRenderer(RendererItem rendererItem) =>
             Native.LibVLCMediaPlayerSetRenderer(NativeReference, rendererItem.NativeReference) == 0;
 
-#region Enums
+#region Events
 
-     
+        readonly object _lock = new object();
 
+        EventHandler<MediaPlayerMediaChangedEventArgs> _mediaPlayerMediaChanged;
+        EventHandler<EventArgs> _mediaPlayerNothingSpecial;
+        EventHandler<EventArgs> _mediaPlayerOpening;
+        EventHandler<MediaPlayerBufferingEventArgs> _mediaPlayerBuffering;
+        EventHandler<EventArgs> _mediaPlayerPlaying;
+        EventHandler<EventArgs> _mediaPlayerPaused;
+        EventHandler<EventArgs> _mediaPlayerStopped;
+        EventHandler<EventArgs> _mediaPlayerForward;
+        EventHandler<EventArgs> _mediaPlayerBackward;
+        EventHandler<EventArgs> _mediaPlayerEndReached;
+        EventHandler<EventArgs> _mediaPlayerEncounteredError;
+        static EventHandler<MediaPlayerTimeChangedEventArgs> _mediaPlayerTimeChanged;
+        EventHandler<MediaPlayerPositionChangedEventArgs> _mediaPlayerPositionChanged;
+        EventHandler<MediaPlayerSeekableChangedEventArgs> _mediaPlayerSeekableChanged;
+        EventHandler<MediaPlayerPausableChangedEventArgs> _mediaPlayerPausableChanged;
+        EventHandler<MediaPlayerTitleChangedEventArgs> _mediaPlayerTitleChanged;
+        EventHandler<MediaPlayerChapterChangedEventArgs> _mediaPlayerChapterChanged; //vlc 3
+        EventHandler<MediaPlayerSnapshotTakenEventArgs> _mediaPlayerSnapshotTaken;
+        EventHandler<MediaPlayerLengthChangedEventArgs> _mediaPlayerLengthChanged;
+        EventHandler<MediaPlayerVoutEventArgs> _mediaPlayerVout;
+        EventHandler<MediaPlayerScrambledChangedEventArgs> _mediaPlayerScrambledChanged;
+        EventHandler<MediaPlayerESAddedEventArgs> _mediaPlayerESAdded; // vlc 3
+        EventHandler<MediaPlayerESDeletedEventArgs> _mediaPlayerESDeleted; // vlc 3
+        EventHandler<MediaPlayerESSelectedEventArgs> _mediaPlayerESSelected; // vlc 3
+        EventHandler<MediaPlayerAudioDeviceEventArgs> _mediaPlayerAudioDevice; // vlc 3
+        EventHandler<EventArgs> _mediaPlayerCorked; // vlc 2.2
+        EventHandler<EventArgs> _mediaPlayerUncorked; // vlc 2.2
+        EventHandler<EventArgs> _mediaPlayerMuted; // vlc 2.2
+        EventHandler<EventArgs> _mediaPlayerUnmuted; // vlc 2.2
+        EventHandler<MediaPlayerVolumeChangedEventArgs> _mediaPlayerVolumeChanged; // vlc 2.2
 
+        public event EventHandler<MediaPlayerMediaChangedEventArgs> MediaChanged
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerMediaChanged += value;
+                    AttachEvent(EventType.MediaPlayerMediaChanged, OnMediaChanged);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerMediaChanged -= value;
+                    DetachEvent(EventType.MediaPlayerMediaChanged, OnMediaChanged);
+                }
+            }
+        }
+
+        public event EventHandler<EventArgs> NothingSpecial
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerNothingSpecial += value;
+                    AttachEvent(EventType.MediaPlayerNothingSpecial, OnNothingSpecial);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerNothingSpecial -= value;
+                    DetachEvent(EventType.MediaPlayerNothingSpecial, OnNothingSpecial);
+                }
+            }
+        }
+
+        public event EventHandler<EventArgs> Opening
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerOpening += value;
+                    AttachEvent(EventType.MediaPlayerOpening, OnOpening);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerOpening -= value;
+                    DetachEvent(EventType.MediaPlayerOpening, OnOpening);
+                }
+            }
+        }
+
+        public event EventHandler<MediaPlayerBufferingEventArgs> Buffering
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerBuffering += value;
+                    AttachEvent(EventType.MediaPlayerBuffering, OnBuffering);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerBuffering -= value;
+                    DetachEvent(EventType.MediaPlayerBuffering, OnBuffering);
+                }
+            }
+        }
+
+        public event EventHandler<EventArgs> Playing
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerPlaying += value;
+                    AttachEvent(EventType.MediaPlayerPlaying, OnPlaying);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerPlaying -= value;
+                    DetachEvent(EventType.MediaPlayerPlaying, OnPlaying);
+                }
+            }
+        }
+
+        public event EventHandler<EventArgs> Paused
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerPaused += value;
+                    AttachEvent(EventType.MediaPlayerPaused, OnPaused);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerPaused -= value;
+                    DetachEvent(EventType.MediaPlayerPaused, OnPaused);
+                }
+            }
+        }
+
+        public event EventHandler<EventArgs> Stopped
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerStopped += value;
+                    AttachEvent(EventType.MediaPlayerStopped, OnStopped);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerStopped -= value;
+                    DetachEvent(EventType.MediaPlayerStopped, OnStopped);
+                }
+            }
+        }
+
+        public event EventHandler<EventArgs> Forward
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerForward += value;
+                    AttachEvent(EventType.MediaPlayerForward, OnForward);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerForward -= value;
+                    DetachEvent(EventType.MediaPlayerForward, OnForward);
+                }
+            }
+        }
+
+        public event EventHandler<EventArgs> Backward
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerBackward += value;
+                    AttachEvent(EventType.MediaPlayerBackward, OnBackward);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerBackward -= value;
+                    DetachEvent(EventType.MediaPlayerBackward, OnBackward);
+                }
+            }
+        }
+
+        public event EventHandler<EventArgs> EndReached
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerEndReached += value;
+                    AttachEvent(EventType.MediaPlayerEndReached, OnEndReached);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerEndReached -= value;
+                    DetachEvent(EventType.MediaPlayerEndReached, OnEndReached);
+                }
+            }
+        }
+
+        public event EventHandler<EventArgs> EncounteredError
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerEncounteredError += value;
+                    AttachEvent(EventType.MediaPlayerEncounteredError, OnEncounteredError);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerEncounteredError -= value;
+                    DetachEvent(EventType.MediaPlayerEncounteredError, OnEncounteredError);
+                }
+            }
+        }
+
+        public event EventHandler<MediaPlayerTimeChangedEventArgs> TimeChanged
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerTimeChanged += value;
+                    AttachEvent(EventType.MediaPlayerTimeChanged, OnTimeChanged);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerTimeChanged -= value;
+                    DetachEvent(EventType.MediaPlayerTimeChanged, OnTimeChanged);
+                }
+            }
+        }
+
+        public event EventHandler<MediaPlayerPositionChangedEventArgs> PositionChanged
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerPositionChanged += value;
+                    AttachEvent(EventType.MediaPlayerPositionChanged, OnPositionChanged);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerPositionChanged -= value;
+                    DetachEvent(EventType.MediaPlayerPositionChanged, OnPositionChanged);
+                }
+            }
+        }
+
+        public event EventHandler<MediaPlayerSeekableChangedEventArgs> SeekableChanged
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerSeekableChanged += value;
+                    AttachEvent(EventType.MediaPlayerSeekableChanged, OnSeekableChanged);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerSeekableChanged -= value;
+                    DetachEvent(EventType.MediaPlayerSeekableChanged, OnSeekableChanged);
+                }
+            }
+        }
+
+        public event EventHandler<MediaPlayerPausableChangedEventArgs> PausableChanged
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerPausableChanged += value;
+                    AttachEvent(EventType.MediaPlayerPausableChanged, OnPausableChanged);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerPausableChanged -= value;
+                    DetachEvent(EventType.MediaPlayerPausableChanged, OnPausableChanged);
+                }
+            }
+        }
+
+        public event EventHandler<MediaPlayerTitleChangedEventArgs> TitleChanged
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerTitleChanged += value;
+                    AttachEvent(EventType.MediaPlayerTitleChanged, OnTitleChanged);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerTitleChanged -= value;
+                    DetachEvent(EventType.MediaPlayerTitleChanged, OnTitleChanged);
+                }
+            }
+        }
+
+        public event EventHandler<MediaPlayerChapterChangedEventArgs> ChapterChanged
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerChapterChanged += value;
+                    AttachEvent(EventType.MediaPlayerChapterChanged, OnChapterChanged);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerChapterChanged -= value;
+                    DetachEvent(EventType.MediaPlayerChapterChanged, OnChapterChanged);
+                }
+            }
+        }
+
+        public event EventHandler<MediaPlayerSnapshotTakenEventArgs> SnapshotTaken
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerSnapshotTaken += value;
+                    AttachEvent(EventType.MediaPlayerSnapshotTaken, OnSnapshotTaken);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerSnapshotTaken -= value;
+                    DetachEvent(EventType.MediaPlayerSnapshotTaken, OnSnapshotTaken);
+                }
+            }
+        }
+
+        public event EventHandler<MediaPlayerLengthChangedEventArgs> LengthChanged
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerLengthChanged += value;
+                    AttachEvent(EventType.MediaPlayerLengthChanged, OnLengthChanged);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerLengthChanged -= value;
+                    DetachEvent(EventType.MediaPlayerLengthChanged, OnLengthChanged);
+                }
+            }
+        }
+
+        public event EventHandler<MediaPlayerVoutEventArgs> Vout
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerVout += value;
+                    AttachEvent(EventType.MediaPlayerVout, OnVout);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerVout -= value;
+                    DetachEvent(EventType.MediaPlayerVout, OnVout);
+                }
+            }
+        }
+
+        public event EventHandler<MediaPlayerScrambledChangedEventArgs> ScrambledChanged
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerScrambledChanged += value;
+                    AttachEvent(EventType.MediaPlayerScrambledChanged, OnScrambledChanged);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerScrambledChanged -= value;
+                    DetachEvent(EventType.MediaPlayerScrambledChanged, OnScrambledChanged);
+                }
+            }
+        }
+
+        // v3
+        public event EventHandler<MediaPlayerESAddedEventArgs> ESAdded
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerESAdded += value;
+                    AttachEvent(EventType.MediaPlayerESAdded, OnESAdded);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerESAdded -= value;
+                    DetachEvent(EventType.MediaPlayerESAdded, OnESAdded);
+                }
+            }
+        }
+
+        // v3
+        public event EventHandler<MediaPlayerESDeletedEventArgs> ESDeleted
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerESDeleted += value;
+                    AttachEvent(EventType.MediaPlayerESDeleted, OnESDeleted);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerESDeleted -= value;
+                    DetachEvent(EventType.MediaPlayerESDeleted, OnESDeleted);
+                }
+            }
+        }
+
+        // v3
+        public event EventHandler<MediaPlayerESSelectedEventArgs> ESSelected
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerESSelected += value;
+                    AttachEvent(EventType.MediaPlayerESSelected, OnESSelected);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerESSelected -= value;
+                    DetachEvent(EventType.MediaPlayerESSelected, OnESSelected);
+                }
+            }
+        }
+
+        // v3
+        public event EventHandler<MediaPlayerAudioDeviceEventArgs> AudioDevice
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerAudioDevice += value;
+                    AttachEvent(EventType.MediaPlayerAudioDevice, OnAudioDevice);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerAudioDevice -= value;
+                    DetachEvent(EventType.MediaPlayerAudioDevice, OnAudioDevice);
+                }
+            }
+        }
+
+        // v2.2
+        public event EventHandler<EventArgs> Corked
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerCorked += value;
+                    AttachEvent(EventType.MediaPlayerCorked, OnCorked);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerCorked -= value;
+                    DetachEvent(EventType.MediaPlayerCorked, OnCorked);
+                }
+            }
+        }
+
+        // v2.2
+        public event EventHandler<EventArgs> Uncorked
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerUncorked += value;
+                    AttachEvent(EventType.MediaPlayerUncorked, OnUncorked);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerUncorked -= value;
+                    DetachEvent(EventType.MediaPlayerUncorked, OnUncorked);
+                }
+            }
+        }
+
+        // v2.2
+        public event EventHandler<EventArgs> Muted
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerMuted += value;
+                    AttachEvent(EventType.MediaPlayerMuted, OnMuted);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerMuted -= value;
+                    DetachEvent(EventType.MediaPlayerMuted, OnMuted);
+                }
+            }
+        }
+
+        // v2.2
+        public event EventHandler<EventArgs> Unmuted
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerUnmuted += value;
+                    AttachEvent(EventType.MediaPlayerUnmuted, OnUnmuted);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerUnmuted -= value;
+                    DetachEvent(EventType.MediaPlayerUnmuted, OnUnmuted);
+                }
+            }
+        }
+
+        // v2.2
+        public event EventHandler<MediaPlayerVolumeChangedEventArgs> VolumeChanged
+        {
+            add
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerVolumeChanged += value;
+                    AttachEvent(EventType.MediaPlayerAudioVolume, OnVolumeChanged);
+                }
+            }
+            remove
+            {
+                lock (_lock)
+                {
+                    _mediaPlayerVolumeChanged -= value;
+                    DetachEvent(EventType.MediaPlayerAudioVolume, OnVolumeChanged);
+                }
+            }
+        }
+
+        void OnMediaChanged(IntPtr ptr)
+        {
+            _mediaPlayerMediaChanged?.Invoke(this,
+                new MediaPlayerMediaChangedEventArgs(RetrieveEvent(ptr).Union.MediaPlayerMediaChanged.NewMedia));
+        }
+
+        void OnNothingSpecial(IntPtr ptr)
+        {
+            _mediaPlayerNothingSpecial?.Invoke(this, EventArgs.Empty);
+        }
+
+        void OnOpening(IntPtr ptr)
+        {
+            _mediaPlayerOpening?.Invoke(this, EventArgs.Empty);
+        }
+
+        void OnBuffering(IntPtr ptr)
+        {
+            _mediaPlayerBuffering?.Invoke(this,
+                new MediaPlayerBufferingEventArgs(RetrieveEvent(ptr).Union.MediaPlayerBuffering.NewCache));
+        }
+
+        void OnPlaying(IntPtr ptr)
+        {
+            _mediaPlayerPlaying?.Invoke(this, EventArgs.Empty);
+        }
+
+        void OnPaused(IntPtr ptr)
+        {
+            _mediaPlayerPaused?.Invoke(this, EventArgs.Empty);
+        }
+
+        void OnStopped(IntPtr ptr)
+        {
+            _mediaPlayerStopped?.Invoke(this, EventArgs.Empty);
+        }
+
+        void OnForward(IntPtr ptr)
+        {
+            _mediaPlayerForward?.Invoke(this, EventArgs.Empty);
+        }
+
+        void OnBackward(IntPtr ptr)
+        {
+            _mediaPlayerBackward?.Invoke(this, EventArgs.Empty);
+        }
+
+        void OnEndReached(IntPtr ptr)
+        {
+            _mediaPlayerEndReached?.Invoke(this, EventArgs.Empty);
+        }
+
+        void OnEncounteredError(IntPtr ptr)
+        {
+            _mediaPlayerEncounteredError?.Invoke(this, EventArgs.Empty);
+        }
+
+        static void OnTimeChanged(IntPtr ptr)
+        {
+            _mediaPlayerTimeChanged?.Invoke(null,
+                new MediaPlayerTimeChangedEventArgs(RetrieveEvent(ptr).Union.MediaPlayerTimeChanged.NewTime));
+        }
+
+        void OnPositionChanged(IntPtr ptr)
+        {
+            _mediaPlayerPositionChanged?.Invoke(this,
+                new MediaPlayerPositionChangedEventArgs(RetrieveEvent(ptr).Union.MediaPlayerPositionChanged.NewPosition));
+        }
+
+        void OnSeekableChanged(IntPtr ptr)
+        {
+            _mediaPlayerSeekableChanged?.Invoke(this,
+                new MediaPlayerSeekableChangedEventArgs(RetrieveEvent(ptr).Union.MediaPlayerSeekableChanged.NewSeekable));
+        }
+
+        void OnPausableChanged(IntPtr ptr)
+        {
+            _mediaPlayerPausableChanged?.Invoke(this,
+                new MediaPlayerPausableChangedEventArgs(RetrieveEvent(ptr).Union.MediaPlayerPausableChanged.NewPausable));
+        }
+
+        void OnTitleChanged(IntPtr ptr)
+        {
+            _mediaPlayerTitleChanged?.Invoke(this,
+                new MediaPlayerTitleChangedEventArgs(RetrieveEvent(ptr).Union.MediaPlayerTitleChanged.NewTitle));
+        }
+
+        void OnChapterChanged(IntPtr ptr)
+        {
+            _mediaPlayerChapterChanged?.Invoke(this,
+                new MediaPlayerChapterChangedEventArgs(RetrieveEvent(ptr).Union.MediaPlayerChapterChanged.NewChapter));
+        }
+
+        void OnSnapshotTaken(IntPtr ptr)
+        {
+            var filenamePtr = RetrieveEvent(ptr).Union.MediaPlayerSnapshotTaken.Filename;
+            var filename = (string)Utf8StringMarshaler.GetInstance().MarshalNativeToManaged(filenamePtr);
+            _mediaPlayerSnapshotTaken?.Invoke(this, new MediaPlayerSnapshotTakenEventArgs(filename));
+        }
+
+        void OnLengthChanged(IntPtr ptr)
+        {
+            _mediaPlayerLengthChanged?.Invoke(this,
+                new MediaPlayerLengthChangedEventArgs(RetrieveEvent(ptr).Union.MediaPlayerLengthChanged.NewLength));
+        }
+
+        void OnVout(IntPtr ptr)
+        {
+            _mediaPlayerVout?.Invoke(this,
+                new MediaPlayerVoutEventArgs(RetrieveEvent(ptr).Union.MediaPlayerVoutChanged.NewCount));
+        }
+
+        void OnScrambledChanged(IntPtr ptr)
+        {
+            _mediaPlayerScrambledChanged?.Invoke(this,
+                new MediaPlayerScrambledChangedEventArgs(RetrieveEvent(ptr).Union.MediaPlayerScrambledChanged.NewScrambled));
+        }
+
+        void OnESAdded(IntPtr ptr)
+        {
+            _mediaPlayerESAdded?.Invoke(this,
+                new MediaPlayerESAddedEventArgs(RetrieveEvent(ptr).Union.EsChanged.Id));
+        }
+
+        void OnESDeleted(IntPtr ptr)
+        {
+            _mediaPlayerESDeleted?.Invoke(this,
+                new MediaPlayerESDeletedEventArgs(RetrieveEvent(ptr).Union.EsChanged.Id));
+        }
+
+        void OnESSelected(IntPtr ptr)
+        {
+            _mediaPlayerESSelected?.Invoke(this,
+                new MediaPlayerESSelectedEventArgs(RetrieveEvent(ptr).Union.EsChanged.Id));
+        }
+
+        void OnAudioDevice(IntPtr ptr)
+        {
+            var deviceNamePtr = RetrieveEvent(ptr).Union.AudioDeviceChanged.Device;
+            var deviceName = (string)Utf8StringMarshaler.GetInstance().MarshalNativeToManaged(deviceNamePtr);
+            _mediaPlayerAudioDevice?.Invoke(this, new MediaPlayerAudioDeviceEventArgs(deviceName));
+        }
+
+        void OnCorked(IntPtr ptr)
+        {
+            _mediaPlayerCorked?.Invoke(this, EventArgs.Empty);
+        }
+
+        void OnUncorked(IntPtr ptr)
+        {
+            _mediaPlayerUncorked?.Invoke(this, EventArgs.Empty);
+        }
+
+        void OnMuted(IntPtr ptr)
+        {
+            _mediaPlayerMuted?.Invoke(this, EventArgs.Empty);
+        }
+
+        void OnUnmuted(IntPtr ptr)
+        {
+            _mediaPlayerUnmuted?.Invoke(this, EventArgs.Empty);
+        }
+
+        void OnVolumeChanged(IntPtr ptr)
+        {
+            _mediaPlayerVolumeChanged?.Invoke(this,
+                new MediaPlayerVolumeChangedEventArgs(RetrieveEvent(ptr).Union.MediaPlayerVolumeChanged.Volume));
+        }
+    
 #endregion
 
 #region Callbacks
